fix(favorite): favorite is out of status bar
