--- src/app/packages/core/components/overlay-geo-registration-error/overlay-geo-registration-error.component.html
@@ -1,4 +1,0 @@
-<div *ngIf="showError" class="geo-registrated">
-	Note: This Image has no Geo-Registration
-</div>
-
