--- src/app/packages/core/components/overlay-geo-registration-error/overlay-geo-registration-error.component.less
@@ -1,7 +1,0 @@
-div.geo-registrated {
-	display: flex;
-	position: absolute;
-	right: 5px;
-	color: #000;
-	background: red;
-}
