--- src/app/packages/core/components/overlay-geo-registration-error/overlay-geo-registration-error.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { OverlayGeoRegistrationErrorComponent } from './overlay-geo-registration-error.component';
-
-describe('OverlayGeoRegistrationErrorComponent', () => {
-	let component: OverlayGeoRegistrationErrorComponent;
-	let fixture: ComponentFixture<OverlayGeoRegistrationErrorComponent>;
-
-	beforeEach(async(() => {
-		TestBed.configureTestingModule({
-			declarations: [OverlayGeoRegistrationErrorComponent]
-		})
-			.compileComponents();
-	}));
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(OverlayGeoRegistrationErrorComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it('should be created', () => {
-		expect(component).toBeTruthy();
-	});
-});
