--- src/app/packages/core/components/overlay-geo-registration-error/overlay-geo-registration-error.component.ts
@@ -1,10 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-	selector: 'ansyn-overlay-geo-registration-error',
-	templateUrl: './overlay-geo-registration-error.component.html',
-	styleUrls: ['./overlay-geo-registration-error.component.less']
-})
-export class OverlayGeoRegistrationErrorComponent {
-	@Input() showError: boolean;
-}
