+++ src/app/packages/status-bar/components/status-bar/status-bar.component.less
@@ -122,6 +122,11 @@ div.status-bar-container {
 			height: 70%;
 			justify-content: space-between;
 			box-sizing: border-box;
+
+			&:not(.show-separator) {
+				margin-left: auto;
+			}
+
 			&.show-separator {
 				border-left: 2px white solid;
 			}
